requirements.txt
streamlit==1.28.1
pandas==2.1.3
streamlit-autorefresh==1.0.1
faiss-cpu==1.7.4
PyPDF2==3.0.1
fastapi==0.104.1
pydantic==2.5.0
pydantic[email]==2.5.0
uvicorn==0.24.0
python-multipart==0.0.6
boto3==1.34.0
botocore==1.34.0 